@@ -1,7 +1,7 @@
 package com.example.fontreszie;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.TextView;
 
 import com.ljx.view.FontResizeView;
